[FIX] Remove failing 'tty' command and replace with redirection to stderr
The 'tty' command in the command substitution (see https://github.com/mortenterhart/heroku-buildpack-wildfly-postgresql/blob/443b41a45d521cd1e72943bbcdc45378b74c79ef/lib/postgresql_datasource.sh#L158) failed on the Heroku build because it has no TTY device and so the ERR trap was executed in this command substitution causing the WildFly to shutdown.

First, this isn't a problem, but the following JBoss commands require the WildFly server to be running. While this isn't the case, the command fails and causes the build to fail.

The solution for this is to remove the 'tty' command and instead redirect the output of 'tee' to the standard error channel.